+++ lib/src/media_container.dart
@@ -4,7 +4,6 @@ import 'package:photo_view/photo_view.dart';
 import 'package:photo_view/photo_view_gallery.dart';
 import 'package:tweet_ui/models/viewmodels/tweet_vm.dart';
 import 'package:tweet_ui/on_tap_image.dart';
-import 'package:tweet_ui/src/animated_play_button.dart';
 import 'package:tweet_ui/src/tweet_video.dart';
 import 'package:tweet_ui/src/view_mode.dart';
 
@@ -18,7 +17,6 @@ class MediaContainer extends StatefulWidget {
     this.tweetVM,
     this.viewMode, {
     Key? key,
-    this.useVideoPlayer = true,
     this.videoPlayerInitialVolume = 0.0,
     this.videoHighQuality = true,
     this.onTapImage,
@@ -26,7 +24,6 @@ class MediaContainer extends StatefulWidget {
 
   final TweetVM tweetVM;
   final ViewMode viewMode;
-  final bool? useVideoPlayer;
   final bool? videoHighQuality;
   final double? videoPlayerInitialVolume;
 
@@ -34,8 +31,7 @@ class MediaContainer extends StatefulWidget {
   _MediaContainerState createState() => _MediaContainerState();
 }
 
-class _MediaContainerState extends State<MediaContainer>
-    with AutomaticKeepAliveClientMixin, TickerProviderStateMixin {
+class _MediaContainerState extends State<MediaContainer> with AutomaticKeepAliveClientMixin, TickerProviderStateMixin {
   late var hashcode;
 
   @override
@@ -50,57 +46,17 @@ class _MediaContainerState extends State<MediaContainer>
 
     Widget? child;
     if (widget.tweetVM.getDisplayTweet().hasSupportedVideo) {
-      if (widget.useVideoPlayer!) {
-        child = TweetVideo(
-          widget.tweetVM.getDisplayTweet(),
-          initialVolume: widget.videoPlayerInitialVolume,
-          videoHighQuality: widget.videoHighQuality,
-        );
-      } else {
-        child = Stack(
-          children: <Widget>[
-            Image(
-              image: CachedNetworkImageProvider(
-                  widget.tweetVM.getDisplayTweet().videoPlaceholderUrl!),
-              width: MediaQuery.of(context).size.width,
-              fit: BoxFit.cover,
-            ),
-            Positioned.fill(
-              child: Padding(
-                padding: const EdgeInsets.only(
-                  left: 4.0,
-                ),
-                child: widget.tweetVM.getDisplayTweet().hasGif
-                    ? Align(
-                        alignment: Alignment.bottomLeft,
-                        child: Image.asset(
-                          "assets/tw__ic_gif_badge.png",
-                          fit: BoxFit.fitWidth,
-                          package: 'tweet_ui',
-                          height: 16,
-                          width: 16,
-                        ),
-                      )
-                    : Container(),
-              ),
-            ),
-            Positioned.fill(
-              child: Center(
-                child: AnimatedPlayButton(
-                  tweetVM: widget.tweetVM.getDisplayTweet(),
-                ),
-              ),
-            )
-          ],
-        );
-      }
+      child = TweetVideo(
+        widget.tweetVM.getDisplayTweet(),
+        initialVolume: widget.videoPlayerInitialVolume,
+        videoHighQuality: widget.videoHighQuality,
+      );
     } else if (widget.tweetVM.getDisplayTweet().hasPhoto) {
       switch (widget.tweetVM.getDisplayTweet().allPhotos.length) {
         case 1:
           child = AspectRatio(
             aspectRatio: MediaContainer.DEFAULT_ASPECT_RATIO_MEDIA_CONTAINER,
-            child: _buildSinglePhoto(context,
-                widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
+            child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
           );
           break;
         case 2:
@@ -110,13 +66,11 @@ class _MediaContainerState extends State<MediaContainer>
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: <Widget>[
                 Expanded(
-                  child: _buildSinglePhoto(context,
-                      widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
+                  child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
                 ),
                 VerticalDivider(color: Colors.white, width: 1.0),
                 Expanded(
-                  child: _buildSinglePhoto(context,
-                      widget.tweetVM.getDisplayTweet().allPhotos, 1, hashcode),
+                  child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 1, hashcode),
                 )
               ],
             ),
@@ -128,27 +82,18 @@ class _MediaContainerState extends State<MediaContainer>
             child: Row(
               children: <Widget>[
                 Expanded(
-                  child: _buildSinglePhoto(context,
-                      widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
+                  child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 0, hashcode),
                 ),
                 VerticalDivider(color: Colors.white, width: 1.0),
                 Expanded(
                   child: Column(
                     children: <Widget>[
                       Expanded(
-                        child: _buildSinglePhoto(
-                            context,
-                            widget.tweetVM.getDisplayTweet().allPhotos,
-                            1,
-                            hashcode),
+                        child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 1, hashcode),
                       ),
                       Divider(color: Colors.white, height: 1.0),
                       Expanded(
-                        child: _buildSinglePhoto(
-                            context,
-                            widget.tweetVM.getDisplayTweet().allPhotos,
-                            2,
-                            hashcode),
+                        child: _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, 2, hashcode),
                       )
                     ],
                   ),
@@ -171,11 +116,7 @@ class _MediaContainerState extends State<MediaContainer>
                 mainAxisSpacing: 1.0,
               ),
               itemBuilder: (BuildContext context, int index) {
-                return _buildSinglePhoto(
-                    context,
-                    widget.tweetVM.getDisplayTweet().allPhotos,
-                    index,
-                    hashcode);
+                return _buildSinglePhoto(context, widget.tweetVM.getDisplayTweet().allPhotos, index, hashcode);
               },
             ),
             aspectRatio: widget.viewMode == ViewMode.standard
@@ -221,8 +162,7 @@ class _MediaContainerState extends State<MediaContainer>
 
   /// allPhotos - list on URLs that is used to build a gallery view
   /// hashcode - used for the Hero tag. The image URL is not enough - for example, you can still have duplicated tweets on a list.
-  Widget _buildSinglePhoto(BuildContext context, List<String> allPhotos,
-      int photoIndex, String hashcode) {
+  Widget _buildSinglePhoto(BuildContext context, List<String> allPhotos, int photoIndex, String hashcode) {
     final List<PhotoViewGalleryPageOptions> galleryPageOptions = allPhotos
         .map((photoUrl) => PhotoViewGalleryPageOptions(
               // TODO add option to choose image size (Twitter supports ":medium" ":large" at the end of photoUrl.