+++ lib/src/quote_tweet_view_embedded.dart
@@ -84,7 +84,6 @@ class QuoteTweetViewEmbed extends StatelessWidget {
               MediaContainer(
                 tweetVM,
                 ViewMode.quote,
-                useVideoPlayer: false,
                 onTapImage: onTapImage,
               ),
             ],